Refactor: Move WorldHeatmap component to /components/common directory
Related to #1634
