+++ src/app/components/common/index.ts
@@ -1,9 +1,11 @@
 import { AnalyticsCard } from "./AnalyticsCard";
 import { BarChart } from "./BarChart";
 import { ChainSelector } from "./ChainSelector";
+import { WorldHeatmap } from "./WorldHeatmap";
 
 export {
   AnalyticsCard,
   BarChart,
   ChainSelector,
+  WorldHeatmap,
 }
